sbrk uses SafeLog
